VCS: remove code for old VCS support (#11377)
* VCS: remove code for old VCS support

This PR removes the code for old VCS support. Besides, it touches the
documentation slightly to remove mentions to particular VCS (Mercurial, Bazaar,
SVN). However, it doesn't re-write all parts of the documentation where we are
using "VCS" in a generic way. There should be another PR that re-phrase those
senteces to mention Git directly.

Related:
* https://github.com/readthedocs/readthedocs.org/pull/11147
* https://github.com/readthedocs/readthedocs.org/issues/8840
* https://about.readthedocs.com/blog/2024/02/drop-support-for-subversion-mercurial-bazaar/

* Add migrations

* Update readthedocs/projects/constants.py

Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>

---------

Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>