Build: update `asdf` and its plugins so we can build latest versions
CircleCI is failing when building `build.tools` because Ruby 3.3.2 is not yet
available in the `asdf` version we have in our Ubuntu Docker image.
Updating the version of `asdf` and its plugins allows us to pre-compile these
versions when running the CircleCI job.

This is required by #11386

Once this PR gets merged, we need to restart the CircleCI job.

https://app.circleci.com/pipelines/github/readthedocs/readthedocs-docker-images/294/workflows/3764492d-0fd4-44a3-bb0c-fccb696dc130/jobs/514