Translations: disable `zh` code (#11385)
* Translations: disable `zh` code

The code `zh` is invalid and project shouldn't select it. They should use `zh-*`
instead. This was causing issues when rendering the new dashboard.

This step only disable this language code so new projects cannot select it.
However, we still need to decide what to do with the db and write a data
migration for them. I suppose that most of them are probably spam, tho.

* Add issue to track the TODO

* Migration