Add historial records for `djstripe.models.Subscription`
Some time ago we wanted this in
https://github.com/readthedocs/meta/issues/53#issuecomment-1234010087. It seems
we wanted this to plot "historical number of subscriptions group by `start_date`
and `status`".

This PR adds historial records to that model for now. If we are happy merging
it, we can come back later and plot more questions related with historial data
and project/organization fields.

Closes #53