API: support rebuilding external version builds with build create endpoint (#11407)
* API: support rebuilding external version builds with build create endpoint

On the old dashboard, we are using a hidden form on two views to retry
builds, but the API is a more consistent way to do this. The feature
this endpoint is missing that BuildTriggerMixin has is retrying a build
for a version using the last build commit.

What is different in this method is that we don't give the user an error
if the build isn't the latest build for a version. To the user, this
detail shouldn't matter and this endpoint instead automatically uses the last
build to get the commit to start the new build.

* Add basic test case for external build trigger

* Refactor tests

* Lint