Addons: add a notification on each Sphinx build
Temporary notification added to all builds using Sphinx to build their docs.
This code should be removed after October 7th, when this behavior becomes the
default.

I'm not sure if this is the direction we want to go or not, but I think we want
to communicate this as much as possible.