Docs: remove "Canonical URLs" guide
This guide is pretty old and it's not correct anymore.
Besides, we are now mentioning `$READTHEDOCS_CANONICAL_URL` environment variable
in the https://docs.readthedocs.io/en/stable/canonical-urls.html.

We should create a redirect from the page deleted to that one where we are
already mentioning how to configure Sphinx and MkDocs.

Closes #10246