Add project: skip config step if YAML file is present
This commit checks if the repository contains a `.readthedocs.yaml` testing
different possible names, and skip the "Add a configuration file to your
project" setp if it's present.

It only works for GitHub for now, but we can probably add GitLab as well.

Related https://github.com/readthedocs/ext-theme/issues/445