Add project: improve way to get the `basics` form
We are adding an extra step in Read the Docs for Business to allow users to
select the Organization and Team as the first step.

Because of that, the _first_ step is not always the `basics` form and now we
need to find it first.

Required by https://github.com/readthedocs/readthedocs-corporate/pull/1851