Allauth: don't completely override the `send_email` method (#11526)
We are overriding the whole send_mail method from allauth in order to use our task to send emails, but by doing so is easy to get out of sync with original method from allauth. Instead of overriding the whole method, I'm just creating a wrapper object in render_mail to send the message using our task.