Standardize error template paths (#11494)
* Default error view drop top level errors and split proxito error views

View changes to support just two paths for errors, `errors/dashboard`
and `errors/proxito`.

* Replace one off teapot view

* Move spam template to standard location

* Point proxito views at new errors/proxito path

* Move top level errors to dashboard error templates path

* Add proxito error templates path with links to dashboard templates for now

* Move proxito 404 errors to proxito error path

* Make error view handler more explicit

Instead of using the status code to find the error template in the error
view handler, just specify a template name. This doesn't return the
corresponding HTTP status code with the error from the debug error view.

* Reword DNS exception reason

"Matching DNS record not found" is not as clear as "Domain not found".

* Add missing template

* Update pattern for error view for proxito

Don't try to default the status code in the view, but instead leave this
alone and let the template decide. This allows for 4xx/5xx fallback templates to
state the actual status code.

* Fix URL load order bug