Drop unnecessary CSP directives for gold view (#11605)
* Drop unnecessary CSP directives for gold view

This does not seem needed, as there is no inline script src in
`subscription_detail.html`. It seems like maybe we wrote this for
`subscription_form.html`, which was old.

This conditional was breaking the view for me locally, as we don't have
any CSP directives for `script-src` locally, we only have these in
production. Because of this, there are no other `script-src` exceptions.

* Revert test functionality