Build: don't show listing or detail view if project is spam
Check for `is_show_dashboard_denied` when accessing build listing or detail
pages and don't show the dashboard if the project is marked as spam.

I tried to write some test cases for this, but it's hard since we don't have
`readthedocs-ext` in here. I didn't find any test for similar cases either.
We could probably mock most of that code, but I don't think it's worth.
I tested this locally and it works as expected at least.

Closes https://github.com/readthedocs/readthedocs-ext/issues/554