Search: respect spacing from block elements when indexing (#11658)
HTML tags can be divided in two categories: inline and block elements.
Inline elements do not start on a new line, while block elements start on a new line.
This gives block elements an implicit spacing that is not present in
inline elements. If there are two tags next to each other,
and one of them is a block element, there will be a space between them.
Or if the two tags are inline elements, there will be no space between them.