Remove session fallback cookie
We set the fallback cookie to work around some old browsers rejecting cookies with `SameSite=None`
https://www.chromium.org/updates/same-site/incompatible-clients/.
That note is from 2019, I think we should be fine without supporting those old browsers,
and browsers from supported devices are already patched to support that cookie.

This is adding an additional cookie that takes some space.