Addons: return all active versions on single version project
We always return all the active versions no matter the versioning scheme of the
project. That's how `versions.active` is expected to work.

However, it seems the conditional was below where these versions are calculated.
I also updated the test case that checks for the active versions on single
versions projects.

Closes https://github.com/readthedocs/sphinx_rtd_theme/issues/1613