Adddons: allow injecting an "always live" JavaScript file
We talked about giving users a way to inject a JavaScript file they control
using our Cloudflare Worker infrastructure to allow them manipulate frozen
documentations.

This could be used in different ways to fix bugs or add features to a particular
frozen set of docs or even to all the versions. The user can make usage of API
data to filter by version or not
(e.g. `if (versions.current == "v3.0") { .. do something ...} `)

The script could live in Read the Docs itself using a relative URL, or outside
it, using an absolute URL.