Addons: make default root CSS selector a shared option
Add a "global" shared CSS root selector that is used all across our features:
link previews, docdiff, search, embed API, etc.

Also, put the "documentation tool name" as a top level config as well. With that
information we can define better defaults for CSS selectors and other stuffs in
the future if needed.

Example: if documentation tool is set as Sphinx by the user, we will be able to
use `[role=main] a.internal` for the Link Preview CSS selector.