Move the `DEFAULT_ROOT_SELECTOR` to the frontend
We are returning `addons.options.root_selector: null` and relying on the
frontend to perform some heuristic to detect the doctool and use the correct
root/link selector based on that.

We are keeping the current behavior for search and filetreediff for now.
Eventually, we should shared the same logic to detect the root selector.