File tree diff: migrate feature flag to model field (#11793)
There is no need to use a feature flag for this. We can use a model
field as we are doing for all the other addons.

As we are not exposing this field to the user yet, they can't enable it
by themselves yet. So, we still have the control of it.

This follows the pattern we have for all the other addons.