Development: try `uv` on Docker images
This is a small test to `uv`. It speeds up the creation of the Docker image a
lot in my case and it doesn't require changing anything else -- we keep using
the exact same `docker.txt` file that we generate with `pip-tool`.

Eventually, if we feel comfortable with it, we can use it more broadly in other
places as well.