EmbedAPI: main content based on documentation tool
We need different CSS selectors to find the main content node depending on the
documentation tool used.

This PR adds the `?selector=` attribute to the EmbedAPI to accept a specific
selector for the main content of the page. This attribute will be sent by the
Addons frontend code using the heuristic code we wrote there.

If this specific selector is not sent, we fallback to the default selector's
value.