Remove old footer code (#11836)
Requests to `/_/api/v2/footer_html`:

* 3k in last 30 days in community:
https://dash.cloudflare.com/76f653d51b8df4e9e956eb83a5ff42f4/readthedocs.io/analytics/traffic?path=%2F_%2Fapi%2Fv2%2Ffooter_html&time-window=43200
* 1k in the last 30 days in commercial:
https://dash.cloudflare.com/76f653d51b8df4e9e956eb83a5ff42f4/readthedocs-hosted.com/analytics/traffic?path=%2F_%2Fapi%2Fv2%2Ffooter_html&time-window=43200

The amount of requests is pretty low, so I guess is just spam or
similar. I checked those Host from where the request was performed and
those projects are using addons correctly and not hitting the old footer
endpoint.

The only project that still hits `footer_html` API is
https://dogapi.readthedocs.io/en/latest/. This is because it's pretty
old (last build on 2016) and our CF Worker is not removing old URL of
our js. We can add it to
https://github.com/readthedocs/common/blob/f5a5c301a3fdb4b03d6176a6980c999c219b14a2/packages/addons-inject/index.js#L23-L54
if we want, tho.

Closes readthedocs/ext-theme#464