Default latest and non-stable notifications to off
These feel like really aggressive defaults for something that will show on most versions.
I think this is part of what people are disliking about the new notifications,
and I'm not really sure why we defaulted them on,
but I'd like to stop the UX issues here if we can.