Build: use actual git identifier for READTHEDOCS_GIT_IDENTIFIER (#11875)
A commit can point to more than one branch/tag, so using git
describe/name-rev or similar doesn't work reliably. Since we already
have the tag/branch name for most of the versions (verbose_name), we can
just use that, there are only three special cases:

- LATEST: this is an alias for the default branch
- STABLE: this is an alias for the stable version
- External versions: we don't save the name of the base branch, just the
commit and PR number, so not sure what to return here, I'm returning the
commit for now. Maybe we can update the docs to mention this.

`commit_name` was doing exactly that, except for stable, so I went ahead
and renamed it and refactored it (we were no longer using it) to use it
in our API, so builders have access to it.

Closes https://github.com/readthedocs/readthedocs.org/issues/11662