Addons: expose `readthedocs.resolver.filename` in the API response
This is required to be able to dynamically update the URLs from the versions in
the flyout when the page is changed using `history` object (without reloading /
SPA) as Docusaurus does.

In that scenario we cannot use the HTML META tag because it was injected with
the old value by CF Worker. Note that the CF Worker is not run after the initial
request on SPA.

The addons frontend will use this attribute to update those URLs accordingly
when the URL changes.

Required by:
* https://github.com/readthedocs/addons/pull/504