Git service: depend on the project instead of users
With https://github.com/readthedocs/readthedocs.org/pull/11942,
a GH app is used to interact with the GH API,
a GH app is not directly tied to a user, so this assumption is not valid anymore.

For most of the operations (except for syncing repos), we don't need to know who is the user
that is performing the action, we just need to know the project that is
being affected.