Projects: require users to have admin permissions to link repositories to projects
Our current logic allows any user (or member) to link a public repository to a project,
this was fine since we use the user session to create the webhook and ssh key on .com,
and if they don't have permissions, nothing will happen.

But with https://github.com/readthedocs/readthedocs.org/pull/11942
operations won't be done on behalf of the user, but the installation,
so we need to make sure that the user has permissions to link the
repository to the project.