Add GitHub App OAuth provider (#12020)
Extracted from https://github.com/readthedocs/common/pull/259/. The
provider is not exposed to users yet, and we allow only staff users to
use it (manually going to /accounts/githubapp/login).

There are steps for ops team and Eric or Anthony to do:

- Create a new GH app from
https://github.com/organizations/readthedocs/settings/apps/new (the name
will be used when we do actions as the installation, like when creating
a comment).
- Callback URL should be
https://app.readthedocs.org/accounts/githubapp/login/callback/
- Keep marked "Expire user authorization tokens"
- Don't active the webhook, since we aren't going to use it yet.
- Permissions (can be updated later if required):
- Repository permissions: Commit statuses (read and write, so we can
create commit statuses), Contents (read only, so we can clone repos with
a token), Metadata (read only, so we read the repo collaborators), Pull
requests (read and write, so we can post a comment on PRs in the
future).
- Organization permissions: Members (read only so we can read the
organization members)
- Account permissions: Email addresses (read only, so allauth can fetch
all verified emails)
- Subscribe to events (can be updated later if required): Installation
target, Member, Organization, Membership, Pull request, Push,
Repository.
- Where can this GitHub App be installed?: any account
- Copy the client ID and client secret into ops repo for the githubapp
provider, we can skip setting a webhook secret and private key, as they
won't be used for now.

Same process for the app for .com.