Projects: require users to have admin permissions to link repositories to projects (#12005)
Our current logic allows any member to link a public repository to a
project, this was fine since we use the user session to create the
webhook and ssh key on .com, and if they don't have permissions, nothing
will happen.

But with https://github.com/readthedocs/readthedocs.org/pull/11942
operations won't be done on behalf of the user, but the installation, so
we need to make sure that the user has permissions to link the
repository to the project.

Looks like we already show a warning when a non-admin user is trying to
import a project


https://github.com/readthedocs/ext-theme/blob/b3ed3cf09686d2c0e4c9741bdd926c83c61feab4/readthedocsext/theme/templates/projects/partials/project_create_automatic.html#L179-L179

I guess we should disallow that action instead of telling users that
they will need to manually configure the project, another option would
be to allow import the project, but we won't link it to the remote
repository, so basically a manual import, but the clone and default
branch fields will auto complete.