Don't set secret key on builders (#12095)
Did a test locally, and builders work correctly, any attempt to use the
secret key will raise an error, since it's empty. There is a PR in ops
to remove it from there in production (wish we had settings for builders
in this repo instead... a man can dream).

Also, we don't need the debug toolbar in the builders, since they aren't
exposed as web service.