Add GitHub App service (#12072)
This is the big one :D

- After this PR is merged, we will start keeping in sync installations
and remote repositories, but they won't be exposed to users yet (we can
manually set up our projects to use a remote repo to test things out,
private repos are not supported with this PR).
- All operations from the old integration are supported, except for
cloning private repos (will open another PR for that).
- PyGitHub is an okay solution, it abstracts dealing with the JWT and
has types for everything. But it can also be inefficient in some cases
(like to interact with a commit, you need to retrieve the commit from
the API first), but I think that's fine for now, we can always fallback
to do this manually, and hopefully shouldn't be that much work, or the
other option is keep using PyGitHub and use a raw call
(requester.getJsonResponse("url")).
- Webhooks always recover nicely in case the installation wasn't
tracked, so if one operation fails for whatever reason (a bug, github is
down, our app is down), we can recover the state in the next operation.
Users can also manually hit "sync".
- The webhook is exposed from the `/webhook/githubapp/` URL, we can
change that if needed.
- We need to add the secrets and IDs on the ops repos.
- Could I have divided this into smaller PRs? Maybe, but that may
require more time removing the tests that are not needed...

Extracted from
https://github.com/readthedocs/readthedocs.org/pull/11942/