Add migration page for GitHub App (#12112)
Can be tested together with
https://github.com/readthedocs/ext-theme/pull/570, there is a video
there showing the workflow.

What this does is groups all repositories the user has linked to a
project, and works over that. The actions the user takes are as follows:

- Connect its account to our new GH app, we validate that the new
account matches the account from the old integration.
- Install our app into the organization/users that the user has
repositories imported from. Users that are part of projects where they
don't have access to the repository owner/organization, can omit those,
so the proper owner can migrate those, or in the case of organizations,
they can request for an admin to approve the installation.
- Users can migrate their projects to the new GitHub app only if the app
is installed in the proper repository, and if they have admin access to
the repository.
- The migration basically consists of removing the old webhooks/ssh keys
from the repo as they are no longer needed, and after that changing the
remote repository to the one from the GH app.
- After all that is done, the next step is to revoke access to our old
OAuth app, the user must do this, as we can't do that on behalf of the
user. This step is recommended but not necessary required (we do make it
required in our migration page, so we aren't responsible for lingering
accounts).
- The final step is basically disconnecting the old GH OAuth app from
the social accounts of the user, once that is done, we no longer can
migrate the project on behalf of the user (or at least the part about
removing the ssh key and webhook), connecting the project to the new
remote repository can still be done, but users can also do that.

Extracted from
https://github.com/readthedocs/readthedocs.org/pull/11942/