GitHub App: fix PR builds and import (#12126)
- Update admin to make it more usable for remote repos/orgs
- When implementing the signature of the base service class I forgot to
return a tuple, the tuple contains a boolean and the response, which is
weird, because all other method return only a boolean, and we never use
the response. So I went ahead and removed the need to return a tuple
from all services.
- Allow activating PR previews for projects connected to a GH app from
the form.
- Build from PRs for projects that explicitly have the option enabled,
since now with the GH app we always listen to all events.
- Tests tests tests