Redirect for `/` and `/welcome` URLs (#12163)
We will always redirect users hitting `readthedocs.org` and
`readthedocs.com` to the `app.` version of it at `/welcome`. From there,
we will redirect to the dashboard if logged in or to the website if not.

If the user is logged in:

readthedocs.org -> app.readthedocs.org/welcome -> app.readthedocs.org

If the user is logged out:

readthedocs.org -> app.readthedocs.org/welcome -> about.readthedocs.com

On the other hand, people hitting the dashboard app, we will redirect to
the login page or to the dashboard.

Closes https://github.com/readthedocs/readthedocs-ops/issues/1625