Build: use scoped credentials for interacting with S3 (#12078)
The main difference from the other credentials, is that we now need to
pass an additional "sesstion_token" value (which django storages calls
security_token :_)

This also brings a little more of things to do if you want to actually
test it locally, as there isn't an STS service for minio.

We need to create a role with the proper policy for production as
described in the dev docs, I guess we need to that from the ops repos?
Not sure how to translate that to salt/terraform, I was able to create
things for dev using the UI.

And, there may be a little more of work involved to remove the storage
credentials from the builder settings, since builders are basically a
full working django app, it's expecting to have access to static
resources somewhere... dealing with settings :_

Closes https://github.com/readthedocs/readthedocs-ops/issues/1599