Build: support cloning private repos with token (#12115)
First I wanted to pass the env var just in the clone step, but we don't
allow passing additional env vars once the environment is created, so
it's available in the whole "clone" environment. The access token we
create is read-only, and should be scoped to just one project as well
(waiting on https://github.com/PyGithub/PyGithub/pull/3287).

Once the clone is done, the token is stored in the .git/config file, so
that token isn't always kept secret from the rest of the build like ssh
keys, but since the token is read-only and scoped to the current
project, and temporary (1 hour). It should be fine. Additionally, the
token is only created for private repos, meaning that only people with
explicit access to the repo may be able to extract the token, but again,
since they already have access to the repo, there is no additional
permissions the token is granting to the user (will document this in
https://github.com/readthedocs/readthedocs.org/pull/12114).