Revert "Remove old media/static files"
This reverts commit d179a3da6f01282b3a2bfb3f45d8fe19f991e2ef.

It seems we are referencing to a bunch of these files
from a lot of places and we cannot delete them.

We think we didn't notice the issue because these files
are still in the S3 bucket, but if we move to another bucket
or provider they will be missing and start failing.