GitLab: Use full path for GitLab groups/organizations (#12236)
name has only the last part of the group namespace. For example, for a
group under the `group/subgroup` namespace, path would be `subgroup`,
and full_path would be `group/subgroup`. The slug is a charfield, so
it's okay to have `/`, looks like we aren't using this field for
anything important, so don't think we need to resync all organizations.

ref https://github.com/readthedocs/readthedocs.org/pull/12233