Add support for subproject and per-version sitemaps, and styled main sitemap
This was a little morning project to expand our default sitemap to make
allow projects to create more granular sitemaps. The idea here is that
the main sitemap becomes a `sitemapindex` sitemap that points to
multiple individual sitemap files. These include:

- A sitemap of version URLs (this is the current prod sitemap)
- A sitemap for each version (new)
- A sitemap for each subproject version (new)

The new sitemaps will 404 if the user doesn't output them.

And in addition, this replaces the XML and XML comment approach, which
doesn't render in many browsers, with a super basic XSLT template. The
template isn't necessiarly needed but it's a nicer experience for
browsers.