Add shot-scraper proof of concept and some updated screen shots
This is not super repeatable just yet, but it at least outputs some
screenshots in an automated way. I don't feel we should spend time
making this repeatable or clean just yet. We could also just end up
deleting the scripting part of this and leave the screenshots. But I'll
probably end up generating more screenshots with this method for now to
see if it's possible to build up this pattern.