Update import private repository UI shots and directions (#12255)
There is a bit more rewriting here than I'd hoped, but figured I'd
update this
page once. This also updates the "importing" to "create project"
nomenclature.

I added shot-scraper for automation of screenshots. This is not super
repeatable
just yet, but it at least outputs some screenshots in an automated way.
I don't
feel we should spend time making this repeatable or clean just yet. We
could
also just end up deleting the scripting part of this and leave the
screenshots.
But I'll probably end up generating more screenshots with this method
for now to
see if it's possible to build up this pattern.

- [ ] This does require a redirect to be added afterwards, for the
previous URL

<!-- readthedocs-preview docs start -->
---
:books: Documentation previews :books:

- User's documentation (`docs`):
https://docs--12255.org.readthedocs.build/12255/

<!-- readthedocs-preview docs end -->

<!-- readthedocs-preview dev start -->
- Developer's documentation (`dev`):
https://dev--12255.org.readthedocs.build/12255/

<!-- readthedocs-preview dev end -->