GitHub App: don't sync repositories that the app doesn't have access to (#12272)
When an app is installed into an organization, GH will send events
related to members being added/removed from repositories, even if those
repositories weren't granted access to the app.. so when attempting to
sync the collaborators from those repositories, there was an error from
GH since we don't have access to those. This is mainly a problem with
public repos, as for private repos GH will always return a 404 if the
app doesn't have access.

This isn't breaking anything from our logic, just creating remote
repositories we don't need and filling sentry with errors.

ref
https://read-the-docs.sentry.io/issues/6701666499/events/0ef3cf2bb9554e39bc2dd9b874aef879/events/?project=148442