Save last login method in a cookie
Initial attempt to save the "last login method" into a cookie so we can show a
visual label next to the button to login.

This is only the backend part. We need to write some code in the template to use
this value next.

I'm not happy with the implementation because looks pretty complex. I didn't
find an easier way to do it and I'm open for suggestions.

Required by https://github.com/readthedocs/ext-theme/issues/421