Builds: run a healthcheck command in the background
Inside the Docker container that runs all the build process, we execute a simple
`curl` command to hit our API at `/build/<id>/healthcheck/` every 2 seconds to
communicate the backend the build is healthy.

The backend runs a periodic task every 5 seconds and check for those builds that
haven't had activity in the last 30 seconds and cancel those builds.

Closes https://github.com/readthedocs/readthedocs.org/issues/11870