Analytics: add `status_code` and regiter analytics from client only
We are moving the pageview for 404 pages to the client side, similar to what we
are doing for 200 pages.

This should reduce the load that bots produce when scanning our docs and hitting
a bunch of 404 pages.

The new approach allows us to enable analytics 200/404 for all the projects
without a feature flag.

Closes #12307