Delete outdated warning message
This was introduced in https://github.com/readthedocs/readthedocs.org/commit/ed0d02b91f1bdad565eb6df7acebb7d01efea18f,
likely when the new bitbucket provider was added.

The registry doesn't have Google and SAML listed,
so users with accounts connected to those providers
are shown this warning, which is confusing.

I checked the DB on .org and .com, and there are no users with
unsupported providers.